fix: bugs in order and cart modules

+++ Frontend/src/app/(user)/cart/page.jsx
@@ -1,50 +1,60 @@
-"use client"
-import React, { useState } from 'react'
-import PageBanner from '../../../../components/user/common/PageBanner'
-import CartItems from '../../../../components/user/cart/CartItems'
-import api from '../../../../components/user/common/api'
-import { toast } from 'react-toastify'
-import { useCart } from '../../../../context/cart/CartContext'
-import { useAuth } from '../../../../context/user/AuthContext'
+"use client";
+import React, { useState } from "react";
+import PageBanner from "../../../../components/user/common/PageBanner";
+import CartItems from "../../../../components/user/cart/CartItems";
+import api from "../../../../components/user/common/api";
+import { toast } from "react-toastify";
+import { useCart } from "../../../../context/cart/CartContext";
+import { useAuth } from "../../../../context/user/AuthContext";
 import Swal from "sweetalert2";
 import withReactContent from "sweetalert2-react-content";
 
-
 const page = () => {
-  const [errMessage, setErrMessage] = useState(null)
-  const {cartData, fetchCartData} = useCart()
+  const [errMessage, setErrMessage] = useState(null);
+  const { cartData, fetchCartData } = useCart();
 
-  const {isAuthenticated} = useAuth()
-  
-  const DataLength = cartData?.length || 0
+  const { isAuthenticated } = useAuth();
 
-   const popup = withReactContent(Swal)
+  const DataLength = cartData?.length || 0;
 
-   const haveCartData = !cartData || !Array.isArray(cartData) || cartData.length <= 0
-  
-  const handleRemoveProduct = async(id, product)=>{
+  const popup = withReactContent(Swal);
+
+  const haveCartData =
+    !cartData || !Array.isArray(cartData) || cartData.length <= 0;
+
+  const handleRemoveProduct = async (id, product) => {
     const result = await popup.fire({
-         title: "Are you sure?",
+      title: "Are you sure?",
       text: `Do you really want to remove ${product} product`,
       icon: "warning",
       showCancelButton: true,
       confirmButtonColor: "#3085d6",
       cancelButtonColor: "#d33",
       confirmButtonText: "Yes, Clear",
-      cancelButtonText: "Cancel"
-      })
+      cancelButtonText: "Cancel",
+    });
 
-      if(!result.isConfirmed) return
-     try {
-      const response = await api.delete(`/cart/${id}`)
-      if(response.status === 200){
-        toast.success(`${product} removed successfully`)
-        fetchCartData()
+    if (!result.isConfirmed) return;
+
+    if (!isAuthenticated) {
+      const cartKey = "guestCart";
+      let existingCart = JSON.parse(localStorage.getItem(cartKey) || "{}");
+
+      const filteredItems =  existingCart.filter(item=> item.product._id !== id)
+      localStorage.setItem(cartKey, JSON.stringify(filteredItems))
+      fetchCartData()
+      return;
+    }
+
+    try {
+      const response = await api.delete(`/cart/${id}`);
+      if (response.status === 200) {
+        toast.success(`${product} removed successfully`);
+        fetchCartData();
       }
-     } catch (error) {
-      if(process.env.NODE_ENV === "development") {
-        console.error(error)
-       
+    } catch (error) {
+      if (process.env.NODE_ENV === "development") {
+        console.error(error);
       }
       const message =
         (Array.isArray(error?.response?.data?.errors) &&
@@ -52,43 +62,40 @@ const page = () => {
         error?.response?.data?.message ||
         "Something went wrong";
       setErrMessage(message);
-      toast.error(message)
-     }
-  }
+      toast.error(message);
+    }
+  };
 
-  const handleClearCart = async()=>{
-    
-
+  const handleClearCart = async () => {
     const result = await popup.fire({
-         title: "Are you sure?",
+      title: "Are you sure?",
       text: "Do you really want to clear cart?",
       icon: "warning",
       showCancelButton: true,
       confirmButtonColor: "#3085d6",
       cancelButtonColor: "#d33",
       confirmButtonText: "Yes, Clear",
-      cancelButtonText: "Cancel"
-      })
+      cancelButtonText: "Cancel",
+    });
 
-      if(!result.isConfirmed) return
+    if (!result.isConfirmed) return;
 
-    if(!isAuthenticated){
-      localStorage.clear("guestCart")
-      fetchCartData()
-      toast.success("Cart clear successfully")
-      return
+    if (!isAuthenticated) {
+      localStorage.clear("guestCart");
+      fetchCartData();
+      toast.success("Cart clear successfully");
+      return;
     }
-    
+
     try {
-      const response = await api.delete(`/cart`)
-      if(response.status === 200){
-          toast.success("Cart clear successfully")
-          fetchCartData()
+      const response = await api.delete(`/cart`);
+      if (response.status === 200) {
+        toast.success("Cart clear successfully");
+        fetchCartData();
       }
     } catch (error) {
-      if(process.env.NODE_ENV === "development") {
-        console.error(error)
-       
+      if (process.env.NODE_ENV === "development") {
+        console.error(error);
       }
       const message =
         (Array.isArray(error?.response?.data?.errors) &&
@@ -96,20 +103,21 @@ const page = () => {
         error?.response?.data?.message ||
         "Something went wrong";
       setErrMessage(message);
-      toast.error(message)
-     }
-    
-  }
-
-  
+      toast.error(message);
+    }
+  };
 
   return (
     <>
-      <PageBanner heading={'Cart'}/>
-      <CartItems handleRemoveProduct={handleRemoveProduct} handleClearCart={handleClearCart} DataLength={DataLength} haveCartData={haveCartData}/>
+      <PageBanner heading={"Cart"} />
+      <CartItems
+        handleRemoveProduct={handleRemoveProduct}
+        handleClearCart={handleClearCart}
+        DataLength={DataLength}
+        haveCartData={haveCartData}
+      />
     </>
-  )
-}
+  );
+};
 
-export default page
-
+export default page;
